Added constants for email functionality
Inserted SENDER and EMAILPASS

Sender is a valid email account, and EMAILPASS is a password or google external email code. 
Both are defined at the top of the file and should be altered before usage. Contact Luc Cohen (Me) for credentials in order to use the QRin bot account. 
Contact details can be found on my profile.